feat(proto): widen ReplicaId to u16 across every wire surface
A replica index is a u16, decoupling replica-identity width from the u8
range. Every wire surface a replica id crosses widens with it:

- the message-body codec and each per-variant `encoded_len` replica
  contribution (U8 -> U16; the RequestSync recovery flag and the per-entry
  body-state tag stay u8);
- the two sync carrier overhead constants count the replica as 2 bytes, so
  `max_unchunked_snapshot_len` and `SYNC_CHUNK_LEN` stay byte-exact against
  the frame cap;
- the labeled hello encodes the replica id as 2 big-endian bytes (a replica
  hello is now 21 bytes; HELLO_VERSION bumped);
- the CertOid identity extension accepts a u16 replica id.

The three replica-keyed maps (peer_checkpoint, dvc_from, sync_serving) key
by ReplicaId so no call site truncates an id to u8. WIRE_VERSION bumped and
the affected codec/frame goldens re-pinned; codec, hello, and CertOid round
trips gain coverage for ids above the u8 range.

+++ viewstamp-proto/src/config.rs
@@ -85,7 +85,7 @@ pub enum ConfigError {
   #[error("replica index {index} out of range for a {count}-replica cluster")]
   ReplicaIndexOutOfRange {
     /// The offending replica index.
-    index: u8,
+    index: u16,
     /// The cluster size.
     count: u8,
   },
@@ -140,7 +140,7 @@ impl Config {
     if replica_count == 0 {
       return Err(ConfigError::ZeroReplicaCount);
     }
-    if replica.get() >= replica_count {
+    if replica.get() >= replica_count as u16 {
       return Err(ConfigError::ReplicaIndexOutOfRange {
         index: replica.get(),
         count: replica_count,
@@ -305,7 +305,7 @@ impl Config {
   /// The primary for a given view: `view % replica_count`.
   #[cfg_attr(not(tarpaulin), inline(always))]
   pub const fn primary(&self, view: View) -> ReplicaId {
-    ReplicaId::new((view.get() % self.replica_count as u64) as u8)
+    ReplicaId::new((view.get() % self.replica_count as u64) as u16)
   }
 
   /// Whether this replica is the primary for `view`.
